Finished!

+++ shared/timeutil/slottime.go
@@ -8,7 +8,7 @@ import (
 
 // SlotStartTime returns the start time in terms of its unix epoch
 // value.
-func SlotStartTime(genesis time.Time, slot uint64) time.Time {
+func SlotStartTime(genesis time.Time, slot uint32) time.Time {
 	duration := time.Second * time.Duration(slot*config.GenYangNetConfig().SecondsPerSlot)
 	startTime := genesis.Add(duration)
 	return startTime
@@ -16,16 +16,16 @@ func SlotStartTime(genesis time.Time, slot uint64) time.Time {
 
 // SlotsSinceGenesis returns the number of slots since
 // the provided genesis time.
-func SlotsSinceGenesis(genesis time.Time) uint64 {
+func SlotsSinceGenesis(genesis time.Time) uint32 {
 	if genesis.After(Now()) { // Genesis has not occurred yet.
 		return 0
 	}
-	return uint64(Since(genesis).Seconds()) / config.GenYangNetConfig().SecondsPerSlot
+	return uint32(Since(genesis).Seconds()) / config.GenYangNetConfig().SecondsPerSlot
 }
 
 // EpochsSinceGenesis returns the number of slots since
 // the provided genesis time.
-func EpochsSinceGenesis(genesis time.Time) uint64 {
+func EpochsSinceGenesis(genesis time.Time) uint32 {
 	return SlotsSinceGenesis(genesis) / config.GenYangNetConfig().SecondsPerSlot
 }
 
@@ -38,18 +38,18 @@ func DivideSlotBy(timesPerSlot int64) time.Duration {
 }
 
 // return the start time of the next epoch
-func NextEpochTime(genesis time.Time, slot uint64) time.Time {
+func NextEpochTime(genesis time.Time, slot uint32) time.Time {
 	cfg := config.GenYangNetConfig()
 	slot = (slot/cfg.SlotsPerEpoch + 1) * cfg.SlotsPerEpoch // the start slot of next epoch
 	return SlotStartTime(genesis, slot)
 }
 
 // return the start time of the next slot
-func NextSlotTime(genesis time.Time, slot uint64) time.Time {
+func NextSlotTime(genesis time.Time, slot uint32) time.Time {
 	return SlotStartTime(genesis, slot+1)
 }
 
 // return a deadline for next slot
-func SlotDeadline(genesis time.Time, slot uint64) time.Time {
+func SlotDeadline(genesis time.Time, slot uint32) time.Time {
 	return NextSlotTime(genesis, slot)
 }
